+++ src/main/java/alfio/config/support/PlatformProvider.java
@@ -29,7 +29,7 @@ import static java.util.Optional.ofNullable;
  * <p>
  * Supported:
  * - Openshift : pgsql only
- * - Cloud Foundry: pgsql (elephantdb)
+ * - Cloud Foundry: postgres, mysql (injected)
  * - Heroku
  * - local use with system properties
  */
@@ -84,26 +84,23 @@ public enum PlatformProvider {
     /**
      * Cloud Foundry configuration.
      * see https://docs.cloudfoundry.org/buildpacks/java/spring-service-bindings.html
-     * <p>
-     * We assume that the "ElephantSQL" has already been bound to the application.
+     * We assume that either the "MySql" service or "ElephantSql" have already been bound to the application.
      * Anyway, since we use Spring, the Cloud Foundry engine should replace the "DataSource" bean with the right one.
      */
     CLOUD_FOUNDRY {
+
         @Override
         public String getDriveClassName(Environment env) {
-            return POSTGRESQL_DRIVER;
+            return isMySql(env) ? MYSQL_DRIVER : POSTGRESQL_DRIVER;
         }
 
         @Override
-        public String getUrl(Environment env) {
-            return env.getRequiredProperty("vcap.services.elephantsql.credentials.uri");
-        }
+        public String getUrl(Environment env) { return ""; }
 
         @Override
         public String getUsername(Environment env) {
             return "";
         }
-
 
         @Override
         public String getPassword(Environment env) {
@@ -117,17 +114,30 @@ public enum PlatformProvider {
 
         @Override
         public String getDialect(Environment env) {
-            return PGSQL;
+            return isMySql(env) ? MYSQL : PGSQL;
         }
 
         @Override
         public boolean isHosting(Environment env) {
-            return ofNullable(env.getProperty("VCAP_APPLICATION")).isPresent();
+            //check if json object for services is returned
+            //example payload
+            //{
+            //"staging_env_json": {},
+            //"running_env_json": {},
+            //"system_env_json": {
+            //    "VCAP_SERVICES": {
+            //        "p-mysql": [
+            //        {
+            //            "name": "alfio-db",
+            return env.getProperty("VCAP_SERVICES") != null;
+        }
+
+        private boolean isMySql(Environment env) {
+            return ofNullable(env.getProperty("VCAP_SERVICES")).map(props -> props.contains("mysql://")).orElse(false);
         }
     },
 
     HEROKU {
-
         @Override
         public String getDriveClassName(Environment env) {
             return POSTGRESQL_DRIVER;
@@ -218,6 +228,9 @@ public enum PlatformProvider {
 
     private static final String POSTGRESQL_DRIVER = "org.postgresql.Driver";
     public static final String PGSQL = "PGSQL";
+
+    private static final String MYSQL_DRIVER = "com.mysql.jdbc.Driver";
+    public static final String MYSQL = "MYSQL";
 
 
     public String getDriveClassName(Environment env) {