+++ src/main/java/alfio/manager/CheckInManager.java
@@ -16,16 +16,21 @@
  */
 package alfio.manager;
 
+import alfio.manager.support.CheckInStatus;
+import alfio.manager.support.DefaultCheckInResult;
+import alfio.manager.support.OnSitePaymentResult;
+import alfio.manager.support.TicketAndCheckInResult;
 import alfio.model.Event;
 import alfio.model.FullTicketInfo;
 import alfio.model.Ticket;
 import alfio.model.Ticket.TicketStatus;
 import alfio.repository.EventRepository;
 import alfio.repository.TicketRepository;
-import lombok.Data;
+import alfio.util.MonetaryUtil;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.Validate;
+import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
@@ -33,8 +38,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.Optional;
 
-import static alfio.manager.CheckInManager.CheckInStatus.*;
-import static alfio.manager.CheckInManager.CheckInStatus.OK_READY_TO_BE_CHECKED_IN;
+import static alfio.manager.support.CheckInStatus.*;
 import static alfio.util.OptionalWrapper.optionally;
 
 @Component
@@ -42,9 +46,6 @@ import static alfio.util.OptionalWrapper.optionally;
 @Log4j2
 public class CheckInManager {
 
-    public enum CheckInStatus {
-        EVENT_NOT_FOUND, TICKET_NOT_FOUND, EMPTY_TICKET_CODE, INVALID_TICKET_CODE, INVALID_TICKET_STATE, ALREADY_CHECK_IN, MUST_PAY, OK_READY_TO_BE_CHECKED_IN, SUCCESS
-    }
     private final TicketRepository ticketRepository;
     private final EventRepository eventRepository;
 
@@ -65,6 +66,13 @@ public class CheckInManager {
         ticketRepository.updateTicketStatusWithUUID(uuid, TicketStatus.ACQUIRED.toString());
     }
 
+    public TicketAndCheckInResult confirmOnSitePayment(String eventName, String ticketIdentifier, Optional<String> ticketCode) {
+        return eventRepository.findOptionalByShortName(eventName)
+            .flatMap(e -> confirmOnSitePayment(ticketIdentifier).map((String s) -> Pair.of(s, e)))
+            .map(p -> checkIn(p.getRight().getId(), ticketIdentifier, ticketCode))
+            .orElseGet(() -> new TicketAndCheckInResult(null, new DefaultCheckInResult(CheckInStatus.TICKET_NOT_FOUND, "")));
+    }
+
     public Optional<String> confirmOnSitePayment(String ticketIdentifier) {
         Optional<String> uuid = findAndLockTicket(ticketIdentifier)
             .filter(t -> t.getStatus() == TicketStatus.TO_BE_PAID)
@@ -74,11 +82,15 @@ public class CheckInManager {
         return uuid;
     }
 
+    public TicketAndCheckInResult checkIn(String shortName, String ticketIdentifier, Optional<String> ticketCode) {
+        return eventRepository.findOptionalByShortName(shortName).map(e -> checkIn(e.getId(), ticketIdentifier, ticketCode)).orElseGet(() -> new TicketAndCheckInResult(null, new DefaultCheckInResult(CheckInStatus.EVENT_NOT_FOUND, "event not found")));
+    }
+
     public TicketAndCheckInResult checkIn(int eventId, String ticketIdentifier, Optional<String> ticketCode) {
         TicketAndCheckInResult descriptor = extractStatus(eventId, ticketRepository.findByUUIDForUpdate(ticketIdentifier), ticketIdentifier, ticketCode);
         if(descriptor.getResult().getStatus() == OK_READY_TO_BE_CHECKED_IN) {
             checkIn(ticketIdentifier);
-            return new TicketAndCheckInResult(descriptor.getTicket(), new CheckInResult(SUCCESS, "success"));
+            return new TicketAndCheckInResult(descriptor.getTicket(), new DefaultCheckInResult(SUCCESS, "success"));
         }
         return descriptor;
     }
@@ -105,22 +117,29 @@ public class CheckInManager {
     }
 
     public TicketAndCheckInResult evaluateTicketStatus(int eventId, String ticketIdentifier, Optional<String> ticketCode) {
-        return extractStatus(eventId, optionally(() -> ticketRepository.findByUUID(ticketIdentifier)), ticketIdentifier, ticketCode);
+        return extractStatus(optionally(() -> eventRepository.findById(eventId)), optionally(() -> ticketRepository.findByUUID(ticketIdentifier)), ticketIdentifier, ticketCode);
+    }
+
+    public TicketAndCheckInResult evaluateTicketStatus(String eventName, String ticketIdentifier, Optional<String> ticketCode) {
+        return extractStatus(eventRepository.findOptionalByShortName(eventName), optionally(() -> ticketRepository.findByUUID(ticketIdentifier)), ticketIdentifier, ticketCode);
     }
 
     private TicketAndCheckInResult extractStatus(int eventId, Optional<Ticket> maybeTicket, String ticketIdentifier, Optional<String> ticketCode) {
-        Optional<Event> maybeEvent = optionally(() -> eventRepository.findById(eventId));
+        return extractStatus(optionally(() -> eventRepository.findById(eventId)), maybeTicket, ticketIdentifier, ticketCode);
+    }
+
+    private TicketAndCheckInResult extractStatus(Optional<Event> maybeEvent, Optional<Ticket> maybeTicket, String ticketIdentifier, Optional<String> ticketCode) {
 
         if (!maybeEvent.isPresent()) {
-            return new TicketAndCheckInResult(null, new CheckInResult(EVENT_NOT_FOUND, "Event with id " + eventId + " not found"));
+            return new TicketAndCheckInResult(null, new DefaultCheckInResult(EVENT_NOT_FOUND, "Event not found"));
         }
 
         if (!maybeTicket.isPresent()) {
-            return new TicketAndCheckInResult(null, new CheckInResult(TICKET_NOT_FOUND, "Ticket with uuid " + ticketIdentifier + " not found"));
+            return new TicketAndCheckInResult(null, new DefaultCheckInResult(TICKET_NOT_FOUND, "Ticket with uuid " + ticketIdentifier + " not found"));
         }
 
         if(!ticketCode.filter(StringUtils::isNotEmpty).isPresent()) {
-            return new TicketAndCheckInResult(null, new CheckInResult(EMPTY_TICKET_CODE, "Missing ticket code"));
+            return new TicketAndCheckInResult(null, new DefaultCheckInResult(EMPTY_TICKET_CODE, "Missing ticket code"));
         }
 
         Ticket ticket = maybeTicket.get();
@@ -131,35 +150,24 @@ public class CheckInManager {
         log.trace("true code    is {}", ticket.ticketCode(event.getPrivateKey()));
 
         if (!code.equals(ticket.ticketCode(event.getPrivateKey()))) {
-            return new TicketAndCheckInResult(null, new CheckInResult(INVALID_TICKET_CODE, "Ticket qr code does not match"));
+            return new TicketAndCheckInResult(null, new DefaultCheckInResult(INVALID_TICKET_CODE, "Ticket qr code does not match"));
         }
 
         final TicketStatus ticketStatus = ticket.getStatus();
 
         if (ticketStatus == TicketStatus.TO_BE_PAID) {
-            return new TicketAndCheckInResult(ticket, new CheckInResult(MUST_PAY, "Must pay for ticket")); //TODO: must say how much
+            return new TicketAndCheckInResult(ticket, new OnSitePaymentResult(MUST_PAY, "Must pay for ticket", MonetaryUtil.addVAT(ticket.getOriginalPrice(), event.getVat()), event.getCurrency()));
         }
 
         if (ticketStatus == TicketStatus.CHECKED_IN) {
-            return new TicketAndCheckInResult(ticket, new CheckInResult(ALREADY_CHECK_IN, "Error: already checked in"));
+            return new TicketAndCheckInResult(ticket, new DefaultCheckInResult(ALREADY_CHECK_IN, "Error: already checked in"));
         }
 
         if (ticket.getStatus() != TicketStatus.ACQUIRED) {
-            return new TicketAndCheckInResult(ticket, new CheckInResult(INVALID_TICKET_STATE, "Invalid ticket state, expected ACQUIRED state, received " + ticket.getStatus()));
+            return new TicketAndCheckInResult(ticket, new DefaultCheckInResult(INVALID_TICKET_STATE, "Invalid ticket state, expected ACQUIRED state, received " + ticket.getStatus()));
         }
 
-        return new TicketAndCheckInResult(ticket, new CheckInResult(OK_READY_TO_BE_CHECKED_IN, "Ready to be checked in"));
+        return new TicketAndCheckInResult(ticket, new DefaultCheckInResult(OK_READY_TO_BE_CHECKED_IN, "Ready to be checked in"));
     }
 
-    @Data
-    public static class TicketAndCheckInResult {
-        private final Ticket ticket;
-        private final CheckInResult result;
-    }
-
-    @Data
-    public static class CheckInResult {
-        private final CheckInStatus status;
-        private final String message;
-    }
 }